+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## master (unreleased)
+* Adds JQuery install generator. ([@abhaynikam][])
 * Adds Bootstrap install generator. ([@abhaynikam][])
 
 ## 0.1.0 (September 5th, 2020)